MCP 101 Tutorial Build Your Own Modular AI Agent for Stock Investment Insights/requirements.txt
langchain-mcp-adapters==0.1.9
mcp==1.12.3
langgraph==0.6.3
langchain-huggingface==0.3.1
python-dotenv
huggingface-hub
fastmcp